+++ src/components/ButtonComponent.js
@@ -1,0 +1,8 @@
+function Calculator(property) {
+  const { name, className } = property;
+  return (
+    <button className={className} type="submit">{name}</button>
+  );
+}
+
+export default Calculator;
